Added framework target for Objective-C wrapper to support Carthage

+++ Cryptobox/Cryptobox/CBVector+Internal.h
@@ -15,6 +15,6 @@
 
 - (nonnull uint8_t *)dataArray;
 
-- (size_t)length;
+- (unsigned long)length;
 
 @end
